드로잉 추첨 이벤트 테이블 및 추첨 메서드 도메인 변경합니다. (#110)
* feat: 추첨 메서드 도메인 변경 (#108)

* test: 추첨 이벤트 테스트 코드 작성 (#108)

* feat: Import 부분 삭제 (#108)

* feat: 유효성 검사 메서드 추출 (#108)
